Fix PersistentObject copy race condition added in 258959e8.
The locking reduction of 258959e8 was overly aggressive: any deep copy of
the PersistentObject object graph must be done under synchronization.

In other words, it's not sufficient to deep-copy an object graph while not
synchronized, and then synchronize while setting the final root reference,
because all of the indirectly referenced objects in the graph are not
protected and therefore could be read incorrectly by another thread.