Add configurable input queue size limit to ChannelNetwork.
Replace executor for handler notifications with a dedicated thread.

Previously it was possible for the queue of executor notifications
to grow without bound if the handler's processing of input data was
slower than the rate at which that data was being received, causing
the handler to get more and more behind.

Instead, now we stop reading input data when the input queue reaches
its maximum size. This will avoid queueing up stale information and
instead, for TCP connections at least, this should create flow control
pushback to the sender to slow the rate of transmission.