Fix bugs in the new non-reflective MessageFmt code added in b0781429.
Instead of trying to parse the format string subformat patterns, we can just
replace the subformats with null formats, and then parse that pattern string.
This makes the parsing much simpler and possible via regex.

In addition, apply the workaround of JDK-8323699 in Segment.toPattern()
(replacing the previous incorrect one) to prevent that bug from affecting
our MessageFmt.toPattern() method.